Move view for changing user region to addons/osfstorage.
